+++ tests/test_regression_parallel.py
@@ -74,59 +74,61 @@ def test_yaml_execution(yaml_file):
         if check_dict is None:
             errors += f"Missing Group - {module_name}\n"
 
-        ### Get Wildcard Tolerances for module ###
-        wild_tol_keys = []
-        for key in tol_dict.keys():
-            if "*" in key:
-                wild_tol_keys.append(key)
+        else:
 
-        ### Check each value in the module
-        for key, truth_value in truth_dict.items():
-            check_value = check_dict.get(key,None)
-            tol_value = None
+            ### Get Wildcard Tolerances for module ###
+            wild_tol_keys = []
+            for key in tol_dict.keys():
+                if "*" in key:
+                    wild_tol_keys.append(key)
 
-            ### Check if there is a valid wildcard ###
-            use_wildcard = False
-            for wild_key in wild_tol_keys:
-                filtered_wild_key = wild_key.replace('*', '')
-                if filtered_wild_key in key:
-                    wild_tol_value = tol_dict[wild_key]
-                    use_wildcard = True
+            ### Check each value in the module
+            for key, truth_value in truth_dict.items():
+                check_value = check_dict.get(key,None)
+                tol_value = None
 
-            ### Check if the exact key is available ###
-            if key in tol_dict.keys():
-                tol_value = tol_dict[key]
+                ### Check if there is a valid wildcard ###
+                use_wildcard = False
+                for wild_key in wild_tol_keys:
+                    filtered_wild_key = wild_key.replace('*', '')
+                    if filtered_wild_key in key:
+                        wild_tol_value = tol_dict[wild_key]
+                        use_wildcard = True
 
-            ### Check if wildcard tolerance key is available ###
-            elif use_wildcard:
-                tol_value = wild_tol_value
+                ### Check if the exact key is available ###
+                if key in tol_dict.keys():
+                    tol_value = tol_dict[key]
 
-            ### Set the default tolerances for a float ###
-            elif isinstance(check_value,float):
-                tol_value = [1e-4,"absolute"]
+                ### Check if wildcard tolerance key is available ###
+                elif use_wildcard:
+                    tol_value = wild_tol_value
 
-            ### Set the default tolerances for a float ###
-            elif isinstance(check_value,int):
-                tol_value = [0,"absolute"]
+                ### Set the default tolerances for a float ###
+                elif isinstance(check_value,float):
+                    tol_value = [1e-4,"absolute"]
 
-            ### Get tolerance parameters ###
-            tol = float(tol_value[0])
-            check_type = tol_value[1]
+                ### Set the default tolerances for a float ###
+                elif isinstance(check_value,int):
+                    tol_value = [0,"absolute"]
+
+                ### Get tolerance parameters ###
+                tol = float(tol_value[0])
+                check_type = tol_value[1]
 
 
-            if check_value is None:
-                errors += f"Missing Key - {module_name}: {key} \n"
-            else:
-                ### Calculate errors ###
-                abs_error = abs(check_value-truth_value)
-                if truth_value != 0:
-                    rel_error = abs(check_value-truth_value)/truth_value
+                if check_value is None:
+                    errors += f"Missing Key - {module_name}: {key} \n"
+                else:
+                    ### Calculate errors ###
+                    abs_error = abs(check_value-truth_value)
+                    if truth_value != 0:
+                        rel_error = abs(check_value-truth_value)/truth_value
 
-            if check_type == "absolute" and abs_error > tol:
-                errors += f"Value Error - {module_name}: {key} (abs error: {abs_error}, tol: {tol} truth: {truth_value}, check: {check_value})\n"
+                if check_type == "absolute" and abs_error > tol:
+                    errors += f"Value Error - {module_name}: {key} (abs error: {abs_error}, tol: {tol} truth: {truth_value}, check: {check_value})\n"
 
-            elif check_type == "relative" and rel_error > tol:
-                errors += f"Value Error - {module_name}: {key} (rel error: {rel_error}, tol: {tol}, truth: {truth_value}, check: {check_value})\n"
+                elif check_type == "relative" and rel_error > tol:
+                    errors += f"Value Error - {module_name}: {key} (rel error: {rel_error}, tol: {tol}, truth: {truth_value}, check: {check_value})\n"
 
     if len(errors)>0:
         errors = parallel_yaml_name + "\n" + errors